==> analysis/00-make-grid.Rmd <==
---
title: "Create filled in prediction grid with RCAs excluded"  
author:
- Philina A. English
date: "30 July 2021"
---

#Purpose  

This code is to create a filled in prediction grid for southern outside Hard Bottom Longline survey area. The prediction grid is a 2 km X 2 km grid that extends from 20m to the 1000m bathymetry within this zone and excludes RCAs. 

```{r setup, include = FALSE, cache=FALSE}
# knitr::opts_chunk$set(
#   collapse = TRUE,
#   comment = "#>",
#   echo = FALSE, message = FALSE, warning = FALSE, evaluate = FALSE, cache = TRUE
# )
```


```{r packages, include = FALSE, evaluate = FALSE}
library(sf)
library(sp) #coverting to UTMS
library(rgdal) #converting to UTMS
library(PBSmapping) #depth contours
library(maptools)
library(rgeos)
library(here)
library(tidyverse)
library(raster)
# install.packages("marmap") # needs to be installed but not loaded 
```

To create the grid, I first created bounding boxes to accommodate the strange shape of the grid more closely. 
The west and east limits were defined by the 1000 m and 20m bathymetry contours respectively [data from `PBSmapping` (Schnute, Boers, and Haigh, 2021)]. 
The prediction grid covered the extent of the bounding box and was a 2 km X 2 km resolution. 
Rockfish Conservation Areas (RCAs) were then removed and area of cells and parts of cells not within RCAs was calculated. 
The final dataframe comprised of the locations of the centroids for each of the grid cells (including partial grid cells).  
Mean depth [data from `marmap`] and proportions of bottom substrate categories [data from Cole Fields] are then calculated using weighted means for each cell. 

```{r eval = F, include = F}
# if correction to archived HBLL S grid coordinates not already done, run this chunck
shape <- st_read(here::here("grids/HBLL-N-S/PHMA_S_GRID.shp"))
# plot(shape)

proj1 =  crs(shape)
proj2 <- "+proj=utm +zone=9 +datum=NAD83 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0  +units=m +no_defs" # I believe this is 3156

hbll_sf <- st_as_sf(shape)

hbll_sf_utm <- st_transform(hbll_sf, crs = 3156)
hbll_sf_utm$geometry <- st_centroid(hbll_sf_utm$geometry) %>%
  # since you want the centroids in a second geometry col:
  st_geometry()

hbll_sf_ll <- st_transform(hbll_sf, crs = 4326)
hbll_sf_ll$geometry <- st_centroid(hbll_sf_ll$geometry) %>%
  # since you want the centroids in a second geometry col:
  st_geometry()

# hbll_sf3$geometry[1]

# plot(st_geometry(hbll_sf))
plot(st_geometry(hbll_sf_ll))
plot(st_geometry(hbll_sf_utm))
plot(st_geometry(hbll_sf_utm), cex = 0.2, add = T)

# to add sf geometry back onto dataframe
sfc_as_cols <- function(x, names = c("x","y")) {
  stopifnot(inherits(x,"sf") #&& inherits(sf::st_geometry(x),"sfc_POINT")
  )
  ret <- sf::st_coordinates(x)
  ret <- tibble::as_tibble(ret)
  stopifnot(length(names) == ncol(ret))
  x <- x[ , !names(x) %in% names]
  ret <- setNames(ret,names)
  dplyr::bind_cols(x,ret)
}

hbll_s_grid_ll <- sfc_as_cols(hbll_sf_ll, c("x", "y"))
st_geometry(hbll_s_grid_ll) <- NULL

hbll_s_grid_utm <- sfc_as_cols(hbll_sf_utm, c("x", "y"))
st_geometry(hbll_s_grid_utm) <- NULL

hbll_grid <- dplyr::select(hbll_s_grid_ll, x, y, LONGITUDE, LATITUDE, DEPTH_M) %>%
  rename(longitude = x, latitude = y, old_lon = LONGITUDE, old_lat = LATITUDE, depth = DEPTH_M)

hbll_grid_utm <- dplyr::select(hbll_s_grid_utm, x, y, LONGITUDE, LATITUDE, DEPTH_M) %>%
  rename(old_lon = LONGITUDE, old_lat = LATITUDE, depth = DEPTH_M) %>% mutate(X = x/ 100000, Y = y/ 100000)

hbll_grid <- left_join(hbll_grid, hbll_grid_utm)

hbll_grid <- mutate(hbll_grid, depth = -depth) %>%
  filter(depth > 0)

hbll_s_grid <- list(grid = hbll_grid, cell_area = 2.0)
# saveRDS(hbll_s_grid, here::here("grids/hbll_s_grid.rds"))
```

Load HBLL S grid
```{r loaddat, include = FALSE}
hbll_s_grid <- readRDS(here::here("grids/hbll_s_grid.rds")) # make sure it has the re-extracted cell coordinates saved in chunk above
```

Use bathymetry contours from PBSmapping to define grid area
```{r bathymetry, include = FALSE }
sp_dat1 <- hbll_s_grid$grid %>% filter(latitude < 50.5) #%>% select(x, y)
sp_dat2 <- hbll_s_grid$grid %>% filter(latitude > 50.5) 

## Create the clipping polygons
CP <- as(raster::extent(
  min(hbll_s_grid$grid$x)-1000, max(hbll_s_grid$grid$x)+1000,
  min(hbll_s_grid$grid$y)-1000, max(hbll_s_grid$grid$y)+1000
  ), "SpatialLines")

CP1 <- as(raster::extent(
  min(sp_dat1$x), max(sp_dat1$x)+40000, #+25000
  min(sp_dat1$y)-30000, max(sp_dat1$y)
  ), "SpatialLines")

CP2 <- as(raster::extent(
  min(sp_dat2$x), max(sp_dat2$x),
  min(sp_dat2$y), max(sp_dat2$y)
  ), "SpatialLines")

crsstring <- "+proj=utm +zone=9 +datum=NAD83 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0  +units=m +no_defs" # i believe this is 3156
crsnum <- 3156

proj4string(CP) <- CRS(crsstring) 
proj4string(CP1) <- CRS(crsstring) 
proj4string(CP2) <- CRS(crsstring) 
CP <- st_as_sf(CP)
CP1 <- st_as_sf(CP1)
CP2 <- st_as_sf(CP2)
CP <- st_transform(CP, crs = crsnum)
CP1 <- st_transform(CP1, crs = crsnum)
CP2 <- st_transform(CP2, crs = crsnum)

#use this polygon and clip to the contours in bcBathymetry
# citation("PBSmapping")
data(bcBathymetry) #bathymetry
contour(bcBathymetry$x, bcBathymetry$y, bcBathymetry$z, col = "pink", method = "edge", vfont = c("sans serif", "plain"))
#,   nlevels = 2)
#make a spatial object
cont <- contourLines(bcBathymetry$x, bcBathymetry$y, bcBathymetry$z, nlevels = 1000) 
# class(cont)
#generate contours
clines <- ContourLines2SLDF(cont)
clines@data[["level"]]
lines <- st_as_sf(clines) #make sf object

# SHALLOW BOUNDARY
c.linesfilter <- lines %>% filter(level == 20) 
st_crs(c.linesfilter) <- CRS('+proj=longlat')
c.linesproj <- st_transform(c.linesfilter, crs = crsnum)
gridarea <- st_bbox(st_intersection(st_geometry(CP1), st_geometry(c.linesproj))) %>% st_point %>% st_sfc(crs=3156)
sa_clip <- st_intersection(st_as_sfc(st_bbox(CP1)),c.linesproj)

# DEEP BOUNDARY
c.linesfilter2 <- lines %>% filter(level == 1000) 
st_crs(c.linesfilter2) <- CRS('+proj=longlat')
c.linesproj2 <- st_transform(c.linesfilter2, crs = crsnum)
gridarea2 <- st_bbox(st_intersection(st_geometry(CP1), st_geometry(c.linesproj2))) %>% st_point %>% st_sfc(crs=3156)
sa_clip2 <- st_intersection(st_as_sfc(st_bbox(CP1)),c.linesproj2)

# SHALLOW BOUNDARY IN NORTH
c.linesfilter <- lines %>% filter(level == 20) 
st_crs(c.linesfilter) <- CRS('+proj=longlat')
c.linesproj <- st_transform(c.linesfilter, crs = crsnum)
gridarea <- st_bbox(st_intersection(st_geometry(CP2), st_geometry(c.linesproj))) %>% st_point %>% st_sfc(crs=3156)
sa_clipN <- st_intersection(st_as_sfc(st_bbox(CP2)),c.linesproj)

# DEEP BOUNDARY IN NORTH
c.linesfilter2 <- lines %>% filter(level == 1000) 
st_crs(c.linesfilter2) <- CRS('+proj=longlat')
c.linesproj2 <- st_transform(c.linesfilter2, crs = crsnum)
gridarea2 <- st_bbox(st_intersection(st_geometry(CP2), st_geometry(c.linesproj2))) %>% st_point %>% st_sfc(crs=3156)
sa_clipN2 <- st_intersection(st_as_sfc(st_bbox(CP2)),c.linesproj2)

ggplot() +
  # geom_sf(data=st_as_sfc(st_bbox(CP1)),fill=NA) +
  geom_sf(data=sa_clip) +
  geom_sf(data=sa_clip2) +
  geom_sf(data=sa_clipN) +
  geom_sf(data=sa_clipN2) +
  theme_classic()

#Cast the linestring to points, append the corner, close the linestring, turn into polygon.
pts_shallow <- sa_clip %>% st_geometry %>% st_cast("POINT")
pts_deep <- sa_clip2 %>% st_geometry %>% st_cast("POINT")
pts_shallowN <- sa_clipN %>% st_geometry %>% st_cast("POINT")
pts_deepN <- sa_clipN2 %>% st_geometry %>% st_cast("POINT")
# 
# st_bbox(pts_shallow) <- st_bbox(CP) 

corner <- st_bbox(CP2)[c('xmin','ymax')] %>% st_point %>% st_sfc(crs=crsnum)%>% st_cast("POINT")

# test <- st_intersection(CP1, pts) %>% st_geometry %>% st_cast("POINT")
sa_lineN <- c(pts_shallowN[1:195], pts_deepN[1:90], 
  corner, pts_shallowN[1]) %>% st_combine %>% st_cast('LINESTRING')
sa_lineS <- c(pts_shallow[1:245], pts_deep[1:212], pts_shallow[1]) %>% st_combine %>% st_cast('LINESTRING')
# sa_polyN <- c(pts_shallowN[1:195], pts_deepN[1:90], corner, 
#   pts_shallowN[1]
#   ) %>% st_combine %>% st_cast('LINESTRING') %>% st_polygonize()
# sa_polyS <- c(pts_shallow[1:218], pts_deep[1:212], pts_shallow[1]
#   ) %>% st_combine %>% st_cast('LINESTRING') %>% st_polygonize()
sa_poly <-  c(
  pts_shallowN[1:195], pts_shallow[1:245],
  pts_deep[1:212], pts_deepN[1:90],
  corner, pts_shallowN[1]
  ) %>% st_combine %>% st_cast('LINESTRING') %>% st_polygonize()

## checking for right cut off
# sa_lineS <- c(pts_shallow[1:245], pts_deep[1:212], pts_shallow[1]) %>% st_combine %>% st_cast('LINESTRING')

ggplot() +
  # geom_sf(data=sa_lineN) +
  # geom_sf(data=sa_lineS) +
  geom_sf(data=sa_poly) +
  # coord_sf(xlim = st_bbox(CP)[c(1,3)], # min & max of x values
  #          ylim = st_bbox(CP)[c(2,4)]) +
  theme_classic()
```

Select the resolution (grid_scale) of the grid you are generating and create cells to fill in grid area.
If wanting the match the original survey design, use 2000 to create 2 x 2 km grid.
```{r}
grid_scale <- 1000 # currently using 2 x 2 km, but I think we could change this...

grid1 <- sf::st_make_grid(x = sa_poly,
  offset = c(-1000, 1000),
  cellsize=c(grid_scale, grid_scale))

keep <- st_intersects(sa_poly, grid1)
grid1 <- grid1[unlist(keep)]

new_grid <- st_as_sf(grid1)

ggplot() +
  # geom_sf(data=sa_poly) +
  # geom_sf(data=grid1) +
  geom_sf(data=new_grid) +
  theme_classic()

saveRDS(new_grid, file = here::here(paste0("grids/hbll_shp_grid_expanded", grid_scale, ".rds"))) 
```


```{r eval=FALSE}
## Add in trawl grid -- turns out it all falls within the above so this step not needed/completed
# trawl <- readRDS(here::here("grids/nd_all_synoptic_new.rds")) %>% 
#   mutate(#X = round(X), Y = round(Y),
#     # X = X / 100, Y = Y / 100,
#     X = X*1000, Y = Y*1000,
#     depth_m = depth
#   ) %>% select(X, Y, depth_m) #%>%
#   # filter(
#   #   X > min(hbll_s_grid$grid$X) & X < max(hbll_s_grid$grid$X) &
#   #     Y > min(hbll_s_grid$grid$Y) & Y < max(hbll_s_grid$grid$Y)
#   # )
# 
# min(trawl$X)
# max(trawl$X)
# 
# trawl_sf <- st_as_sf(trawl, coords = c("X", "Y"), crs = 3156)
# trawl_sf_grid <- sf::st_make_grid(x = trawl_sf,
#   offset = st_bbox(trawl_sf)[c("xmin", "ymin")]+ c(-1000, -1000),
#   cellsize=c(2000, 2000))
# 
# keep <- st_intersects(trawl_sf, trawl_sf_grid)
# trawl_sf_grid <- trawl_sf_grid[unlist(keep)]
# trawl_sf2 <- st_as_sf(trawl_sf_grid)
# trawl_sf2$source_data <- "TRAWL"
# 
# ggplot() +
#   geom_sf(data=new_grid, colour = "red") +
#   geom_sf(data=trawl_sf2, alpha = 0.5, colour ="NA") +
#   theme_classic()

# 
# # grid_utm <- full_join(grid, s_grid_utm, by = c("X", "Y")) %>%
# #   mutate(depth = ifelse(is.na(depth_m), depth, depth_m), 
# #     survey = ifelse(is.na(survey), "TRAWL" #survey2
# #       , survey)) %>%
# #   select(-depth_m, #-survey2, 
# #     -latitude, -longitude)
# # 
# # grid_xy <- grid_utm %>%
# #   select(X, Y) %>%
# #   mutate(X = X * 100, Y = Y * 100)
# # grid_ll <- utm2ll(grid_xy) %>% rename(longitude = X, latitude = Y)
# # 
# # grid_utm <- bind_cols(grid_utm, grid_ll)
# # 
# # grid_sf <- st_as_sf(grid_utm, coords = c("longitude", "latitude"))
# # st_crs(grid_sf) <- 4326 # set the coordinate reference system
# # grid <- sfc_as_cols(grid_sf, c("longitude", "latitude"))
# # st_geometry(grid) <- NULL
# # 
# # saveRDS(grid, file = "data-generated/hybrid_grid.rds") 
```

Remove RCAs from grid
```{r}
# get RCA
rcas <- st_read(here::here("grids/RCA2019/RCA_2019.shp"))
## plot(rcas) # takes an impossibly long time so do not do!
# glimpse(rcas)

rcas_utm <- st_transform(rcas, crs = 3156) # convert to utm proj used for grid
# rcas_ll <- st_transform(rcas, crs = 4326) # if needed in lat lon

cells_in_rcas <- st_intersection(new_grid, rcas_utm)
rcas_in_grid <- st_intersection(rcas_utm, new_grid)

# plot(rcas_in_grid)

# this approach of getting centroids fails... none of cell centroids seem to match
# cells_in_rcas$geometry <- st_centroid(cells_in_rcas) %>% st_geometry()
# coords_rcas <- as.data.frame(st_coordinates(cells_in_rcas$geometry))

#not sure why, but st_union required otherwise df becomes huge!
new_grid_norca <- st_difference(new_grid, st_union(cells_in_rcas))

# new_grid_norca <- st_difference(new_grid, st_union(rcas_in_grid))

# st_is_valid(new_grid)
# st_is_valid(cells_in_rcas)
# class(new_grid)
# class(cells_in_rcas)
# st_erase = function(x, y) st_difference(x, st_union(st_combine(y)))

# new_grid_norca <- st_erase(new_grid, cells_in_rcas)
# rca_keep <- st_intersects(new_grid, rcas_utm)
# cells_in_rcas <- rcas_utm[unlist(rca_keep)]
# cells_in_rcas <- st_as_sf(cells_in_rcas)


# saveRDS(new_grid_norca, here::here("grids/hbll_shp_grid_norca.rds"))
saveRDS(new_grid_norca, here::here(paste0("grids/hbll_shp_grid_norca_expanded", grid_scale, ".rds")))

ggplot() +
  geom_sf(data=new_grid_norca) +
  theme_classic()
```

Make dataframe with UTMs and lat lons
```{r}
new_grid_utm <- new_grid_norca # already in UTMs

new_grid_utm$area <- as.numeric(st_area(new_grid_utm))
new_grid_utm$geometry <- st_centroid(new_grid_utm) %>% st_geometry()
hbll_s_grid_utm <- as.data.frame(st_coordinates(new_grid_utm$geometry))
hbll_s_grid_utm$area <- new_grid_utm$area 


if(grid_scale > 1500){
hbll_grid_utm <- dplyr::select(hbll_s_grid_utm, area, X, Y) %>%
  rename(x = X, y = Y) %>% mutate(X = round(x/1000)/100, Y = round(y/1000)/100)
} else {
hbll_grid_utm <- dplyr::select(hbll_s_grid_utm, area, X, Y) %>%
  rename(x = X, y = Y) %>% mutate(X = round(x/1000, 1)/100, Y = round(y/1000, 1)/100)
}

new_grid_ll <- st_transform(new_grid_norca, crs = 4326)
new_grid_ll$geometry <- st_centroid(new_grid_ll) %>% st_geometry()
hbll_s_grid_ll <- as.data.frame(st_coordinates(new_grid_ll$geometry))
hbll_grid_ll <- dplyr::select(hbll_s_grid_ll, X, Y) %>%
  rename(longitude = X, latitude = Y)

# cells_to_remove <- dplyr::select(coords_rcas, X, Y) %>%
#   rename(x = X, y = Y) %>% mutate(X = round(x/1000)/100, Y = round(y/1000)/100)

new_grid_norca  <- bind_cols(hbll_grid_ll, hbll_grid_utm)# %>% filter(area > 2e+06)

ggplot(new_grid_norca) +
  geom_tile(aes(X,Y, width = sqrt(area)/100000, height = sqrt(area)/100000))+
  theme_classic()

# hbll_grid %>% mutate(X = round(X * 100), Y = round(Y * 100)) %>%
#   mutate(X = (X / 100), Y = (Y / 100)) %>% View()
```

Check if range of old and new grids match after rounding
```{r}
new_grid_norca 

old_grid <- readRDS(here::here("grids/hbll_s_grid.rds"))
old_grid <- old_grid$grid %>% mutate(X = round(X * 100), Y = round(Y * 100)) %>%
  mutate(X = (X / 100), Y = (Y / 100))

range(old_grid$X)
range(new_grid_norca$X)

range(old_grid$Y)
range(new_grid_norca$Y)

if(all(c(range(old_grid$X),range(old_grid$Y)) == c(range(new_grid_norca$X),range(new_grid_norca$Y))) ){
saveRDS(new_grid_norca, here::here(paste0("grids/hbll_s_grid_norca", grid_scale, ".rds")))
}

if(all(
    min(old_grid$X) >= min(new_grid_norca$X),  
    max(old_grid$X) <= max(new_grid_norca$X),
    min(old_grid$Y) >= min(new_grid_norca$Y),
    max(old_grid$Y) <= max(new_grid_norca$Y)
    )){
saveRDS(new_grid_norca, here::here(paste0("grids/hbll_s_grid_norca_expanded", grid_scale, ".rds")))
}
```


# Get depth data for all grid cells

Download depth data for extent of new grid
```{r}
f <- here::here("data/bath_res1_marmap.rds")
if(!file.exists(f)){
# Load package
library(marmap)
new_grid_norca <- readRDS(here::here(paste0("grids/hbll_s_grid_norca.rds")))
# Download bathymetric data and save on disk
bat <- getNOAA.bathy(min(new_grid_norca$longitude)-2, max(new_grid_norca$longitude)+2,
  min(new_grid_norca$latitude)-2, max(new_grid_norca$latitude)+2, res = 1, keep = TRUE) #resolution of the grid, in minutes (default is 4 or > 4 miles)
saveRDS(bat, here::here("data/bath_res1_marmap.rds"))
}
```

```{r}
f <- here::here("data/bath_res1_marmap2.rds")
if(!file.exists(f)){
# Load package
library(marmap)
new_grid_norca <- readRDS(here::here(paste0("grids/hbll_s_grid_norca_expanded", grid_scale, ".rds")))
# Download bathymetric data and save on disk
bat <- getNOAA.bathy(min(new_grid_norca$longitude)-2, max(new_grid_norca$longitude)+2,
  min(new_grid_norca$latitude)-2, max(new_grid_norca$latitude)+2, res = 1, keep = TRUE) #resolution of the grid, in minutes (default is 4 or > 4 miles)
saveRDS(bat, here::here("data/bath_res1_marmap2.rds"))
}
```

Explore depth data retrieved
- chunk just exploratory and not working because track was probably from some example
```{r eval=F}
## Load package
# library(marmap)
# # bat <- readRDS(here::here("data/bath_res1_marmap.rds"))
# bat <- readRDS(here::here("data/bath_res1_marmap2.rds"))
# # Create nice looking color palettes
# blues <- c("lightsteelblue4", "lightsteelblue3", "lightsteelblue2", "lightsteelblue1")
# greys <- c(grey(0.6), grey(0.93), grey(0.99))
# # Plot
# plot(bat, image = TRUE, land = TRUE, lwd = 0.1, bpal = list(c(0, max(bat), greys), c(min(bat), 0, blues)))
# plot(bat, lwd = 0.8, deep = -1000, shallow = -20, step = 20, add = TRUE) # highlight coastline
# 
# # Get depth profile along both tracks and remove positive depths (since random fake values can be on land)
# # path.profile() gets depth value for all cells of the bathymetric grid below the gps tracks
# path1 <- path.profile(track1[,-3], bat) ; path1 <- path1[-path1[,4]>0,]
# path2 <- path.profile(track2[,-3], bat) ; path2 <- path2[-path2[,4]>0,]
# # Get depth values for each gps tracking point
# # get.depth() retrieve depth values only for gps tracking points
# depth1 <- get.depth(bat, track1$lon, track1$lat, locator = FALSE, distance = TRUE) ;
# depth2 <- get.depth(bat, track2$lon, track2$lat, locator = FALSE, distance = TRUE) ;
# # Add depth values to tracks 1 and 2 and remove positive depths
# track1$depth <- depth1$depth ; track1 <- track1[-track1$depth > 0,]
# track2$depth <- depth2$depth ; track2 <- track2[-track2$depth > 0,]
# # Plot
# layout(matrix(c(1, 2, 1, 3, 4, 4), ncol = 2, byrow = TRUE), height = c(1, 1, 1))
# ## Bathymetric map with gps tracks
# plot(bat, land = TRUE, image = TRUE, lwd = 0.2, bpal = list(c(min(bat,na.rm=TRUE), 0, blues), c(0, max(bat, na.rm=TRUE), greys)), ylim=c(-36.5,-35.25))
# plot(bat, deep = 0, shallow = 0, step = 0, lwd = 0.8, add = TRUE)
# lines(track1, col = "brown3")
# lines(track2, col = "darkorange1")
# legend("topright", legend = c("Track 1", "Track 2"), lwd = 1, col = c("brown3", "darkorange1"), pch = 1, pt.cex = 0.5, bg="white")
# ## Depths profiles along gps tracks
# plotProfile(path1, main = "Track 1")
# plotProfile(path2, main = "Track 2")
# ## Depth as a function of time since the deployment of gps tags
# par(mar=c(5,4,1,2))
# plot(track1$time, track1$depth, xlab = "Time (arbitrary unit)", ylab = "Depth (m)", type = "o", cex = 0.5, col = "brown3", xlim = range(c(track1$time, track2$time)), ylim = range(c(track1$depth, track2$depth)))
# lines(track2$time, track2$depth, type = "o", cex = 0.5, col = "darkorange1")
# legend("bottomright", legend = c("Track 1", "Track 2"), lwd = 1, col = c("brown3", "darkorange1"), pch = 1, pt.cex = 0.5)
```

Extract depth for grid cells
```{r}
# make depth raster
# bat <- readRDS(here::here("data/bath_res1_marmap.rds"))
bat <- readRDS(here::here("data/bath_res1_marmap2.rds"))
rast_bat <- marmap::as.raster(bat) 

# change projection to match grid
proj <- "+proj=utm +zone=9 +datum=NAD83 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0  +units=m +no_defs"
rast_bat <- projectRaster(rast_bat, crs = proj) 

rast_bat[rast_bat > 0] <- NA

plot(rast_bat)

rast_bat2 <- marmap::as.raster(bat) 
rast_bat2 <- projectRaster(rast_bat2, crs = proj) 
rast_water_area <- rast_bat2
rast_water_area[rast_bat2 <  0] <- 1
rast_water_area[rast_bat2 >=  0] <- 0

plot(rast_water_area)

# depth for complete cells only
# hbll_grid <- readRDS(here::here("grids/hbll_s_grid_norca.rds"))
hbll_grid <- readRDS(here::here(paste0("grids/hbll_s_grid_norca_expanded", grid_scale, ".rds")))
# grid_scale <- 2000 #scale of prediction grid in m
grid <- raster(
  ext = extent(hbll_grid)+c(-1000, 1000, -1000, 1000), # need to add 1km in all directions to allow coordinates to be the centroids of 2x2km cells
  resolution = grid_scale, crs = CRS(proj))

grid.df <- as.data.frame(grid, xy = TRUE)
grid.df$layer <- NULL
depth <- raster::extract(x=rast_bat, y=grid.df, fun = mean, na.rm = T, exact=T)
extracted_depth <- cbind(grid.df, depth)

hbll_grid_full_cells <- left_join(hbll_grid, extracted_depth)

# depth for partial cells too
new_grid_norca <- readRDS(here::here(paste0("grids/hbll_shp_grid_norca_expanded", grid_scale, ".rds")))
new_grid_norca <- st_as_sf(new_grid_norca)
new_grid_norca <- st_transform(new_grid_norca, crs = 3156)

# # unweighted depths
# get_depth <- raster::extract(x=rast_bat, y=new_grid_norca, na.rm = T, exact=T, sp = T)
# mean_depth <- unlist(lapply(get_depth, function(x) if (!is.null(x)) mean(x, na.rm=TRUE) else NA ))

# weighted depths
# library(raster)
get_depth2 <- raster::extract(x=rast_bat, y=new_grid_norca, na.rm = T, exact=T, weights=T, sp = T)
mean_depth2 <- sapply(get_depth2, function(x) if (!is.null(x)) {sum(apply(x, 1, prod), na.rm = T) / sum(x[,2], na.rm = T)} else NA)

# get water area
get_water_area <- raster::extract(x=rast_water_area, y=new_grid_norca, na.rm = T, weights=T, exact=T, sp = T)
water_area <- sapply(get_water_area, function(x) if (!is.null(x)) {sum(apply(x, 1, prod))} else NA)


all_depths <- bind_cols(hbll_grid_full_cells, #mean_depth = mean_depth, 
  prop_water = water_area,
  wm_depth = mean_depth2) %>% 
  mutate(cell_depth = -depth, #depth_uw = -mean_depth, 
    depth_wm = -wm_depth, depth = -wm_depth) %>% dplyr::select(#-mean_depth, 
      -wm_depth)
# all_depths <- all_depths %>% mutate(X = round(x/1000,1)/100, Y = round(y/1000, 1)/100)

saveRDS(all_depths, here::here(paste0("grids/exrtracted_depths_norca_expanded", grid_scale, ".rds")))


# ggplot(all_depths) + geom_point( aes(cell_depth, depth_uw, colour = area),alpha=0.2) # unweighted
ggplot(all_depths) + geom_point( aes(cell_depth, depth_wm, colour = area), alpha=0.2) # weighted
# weighted seems closer to the full cells only way of doing this...
# ggplot(all_depths) + geom_point(aes(depth_wm, depth_uw, colour = area),alpha=0.2)


ggplot(all_depths) + geom_jitter(aes(area, prop_water*area, colour = log(depth)), alpha=0.2) # weighted

# will apply this filter after extracting substrate
hbll_grid_keep <- all_depths %>% 
  # filter(depth_all < 1100) %>% filter(depth_all > 10) 
  filter(depth > 0) 

# ggplot(all_depths) +
ggplot(hbll_grid_keep) +
  geom_point(aes(x,y, colour = log(depth)), alpha = 0.75) +
  scale_colour_viridis_c(direction = -1) +
  # geom_point(aes(X,Y, colour = -cell_depth), alpha = 0.2) +
  theme_classic()
```

# Get substrate covariates

Convert tif projection and save as high res rasters
```{r eval = F}
# f <- "data/highres-sandy-raster.grd" 
f <- here::here("data/highres-sandy-raster-expanded.grd") 
if (!file.exists(f)) {
# retrieve substate layers if not done so before
rocky <- raster(here::here("data/substrate/rocky.tif"))
mixed <- raster(here::here("data/substrate/mixed.tif"))
muddy <- raster(here::here("data/substrate/muddy.tif"))
sandy <- raster(here::here("data/substrate/sandy.tif"))

# original projection
proj <- "+proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0"
projdefs <- "+datum=NAD83 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
geoCRS <- paste( proj, projdefs, sep=" " )

# new projection
proj <- "+proj=utm +zone=9 +datum=NAD83 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0  +units=m +no_defs"

new_rocky <- raster::projectExtent(rocky, crs = proj)
raster::res(new_rocky) <- 100
new_mixed <- raster::projectExtent(mixed, crs = proj)
raster::res(new_mixed) <- 100
new_muddy <- raster::projectExtent(muddy, crs = proj)
raster::res(new_muddy) <- 100
new_sandy <- raster::projectExtent(sandy, crs = proj)
raster::res(new_sandy) <- 100

# # Project values to new raster
hres_rocky <- raster::projectRaster(rocky, new_rocky) #, method = "ngb")
# writeRaster(hres_rocky, file = "data/highres-rocky-raster.grd")
hres_mixed <- raster::projectRaster(mixed, new_mixed) #, method = "ngb")
# writeRaster(hres_mixed, file = "data/highres-mixed-raster.grd")
hres_muddy <- raster::projectRaster(muddy, new_muddy) #, method = "ngb")
# writeRaster(hres_muddy, file = "data/highres-muddy-raster.grd")
hres_sandy <- raster::projectRaster(sandy, new_sandy) #, method = "ngb")
# writeRaster(hres_sandy, file = "data/highres-sandy-raster.grd")
}
```

Extract variables for new grid
```{r}
# load previously saved raster
rocky <- raster(here::here("data/highres-rocky-raster.grd"))
## with weights
get_rocky2 <- raster::extract(x=rocky, y=new_grid_norca, na.rm = T, exact=T, weights=T, sp = T)
mean_rocky2 <- sapply(get_rocky2, function(x) if (!is.null(x)) {sum(apply(x, 1, prod), na.rm = T) / sum(x[,2], na.rm = T)} else NA)
## without weights
# get_rocky <- raster::extract(x=rocky, y=new_grid_norca, na.rm = T, exact=T, sp = T)
# mean_rocky <- unlist(lapply(get_rocky, function(x) if (!is.null(x)) mean(x, na.rm=TRUE) else NA ))

muddy <- raster(here::here("data/highres-muddy-raster.grd"))
## with weights
get_muddy <- raster::extract(x=muddy, y=new_grid_norca, na.rm = T, exact=T, weights=T, sp = T)
mean_muddy <- sapply(get_muddy, function(x) if (!is.null(x)) {sum(apply(x, 1, prod), na.rm = T) / sum(x[,2], na.rm = T)} else NA)
## without weights
# get_muddy <- raster::extract(x=muddy, y=new_grid_norca, na.rm = T, exact=T, sp = T)
# mean_muddy <- unlist(lapply(get_muddy, function(x) if (!is.null(x)) mean(x, na.rm=TRUE) else NA ))

mixed <- raster(here::here("data/highres-mixed-raster.grd"))
get_mixed <- raster::extract(x=mixed, y=new_grid_norca, na.rm = T, exact=T, weights=T, sp = T)
mean_mixed <-  sapply(get_mixed, function(x) if (!is.null(x)) {sum(apply(x, 1, prod), na.rm = T) / sum(x[,2], na.rm = T)} else NA)
## or without weights
# get_mixed <- raster::extract(x=mixed, y=new_grid_norca, na.rm = T, exact=T, sp = T)
# mean_mixed <- unlist(lapply(get_mixed, function(x) if (!is.null(x)) mean(x, na.rm=TRUE) else NA ))

sandy <- raster(here::here("data/highres-sandy-raster.grd"))
get_sandy <- raster::extract(x=sandy, y=new_grid_norca, na.rm = T, exact=T, weights=T, sp = T)
mean_sandy <-  sapply(get_sandy, function(x) if (!is.null(x)) {sum(apply(x, 1, prod), na.rm = T) / sum(x[,2], na.rm = T)} else NA)

# add substrate to grid with depth data
all_covs <- bind_cols(all_depths, rocky = mean_rocky2, #rocky_uw = mean_rocky, 
  muddy = mean_muddy, mixed = mean_mixed, sandy = mean_sandy) 

# # check how weighting changes things?
# ggplot(all_covs) +
#   geom_point(aes(rocky, rocky_uw, colour = log(area)), alpha = 0.25) +
#   theme_classic()

saveRDS(all_covs, here::here(paste0("grids/norca_grid_w_substrate_expanded", grid_scale, ".rds")))
```

Correct the area estimates by the proportion water
```{r}

# depths_only <- readRDS(here::here("grids/exrtracted_depths_norca_expanded.rds"))
# 
# all_covs <- readRDS(here::here("grids/norca_grid_w_substrate_expanded.rds")) %>% 
#   rename(old_depth = depth) %>% dplyr::select( -depth_wm, -cell_depth) %>%  left_join(depths_only)

all_covs2 <- all_covs %>% mutate(cell_area = area, area = area * prop_water) %>% filter(area > 0)
saveRDS(all_covs2, here::here(paste0("grids/norca_grid_w_substrate_expanded", grid_scale, ".rds")))
```


```{r}
ggplot(all_covs) +
  geom_point(aes(x,y, colour = prop_water), alpha = 0.75) +
  scale_colour_viridis_c(direction = -1) +
  # geom_point(aes(X,Y, colour = -cell_depth), alpha = 0.2) +
  theme_classic()

ggplot(all_covs2) +
  geom_point(aes(area, prop_water), alpha = 0.75) +
  scale_colour_viridis_c(direction = -1) +
  # geom_point(aes(X,Y, colour = -cell_depth), alpha = 0.2) +
  theme_classic()

```
